+++ GameShop/src/main/java/com/mcgill/ecse321/GameShop/model/Game.java
@@ -45,7 +45,7 @@ public class Game {
   private int game_id;
   private String title;
   private String description;
-  private int price;
+  private double price;
 
   @Enumerated(EnumType.STRING)
   private GameStatus gameStatus;
@@ -166,7 +166,7 @@ public class Game {
     return description;
   }
 
-  public int getPrice() {
+  public double getPrice() {
     return price;
   }
 
@@ -218,6 +218,7 @@ public class Game {
     List<Platform> newPlatforms = Collections.unmodifiableList(platforms);
     return newPlatforms;
   }
+
   public void setPlatforms(List<Platform> platforms) {
     this.platforms = platforms;
   }
@@ -338,6 +339,7 @@ public class Game {
     }
     return wasAdded;
   }
+
   public void setCategories(List<Category> categories) {
     this.categories = categories;
   }
